refactor(guard): split package execution context
